add calibration with reference image

+++ src/main/resources/python_scripts/manualCalib.py
@@ -13,4 +13,4 @@ raw_phasor_arr = raw_phasor_arr.values.transpose(1, 2, 0).copy()
 
 phasor = phasor_transform(raw_phasor_arr[:, :, 1], raw_phasor_arr[:, :, 2], phase_shift, mod_factor)
 
-output = ij.py.to_dataset(np.array([phasor[0], phasor[1]]))
+output = ij.py.to_dataset(np.array([phasor[0], phasor[1]], dtype=np.float32))
