+++ src/main/resources/python_scripts/phasor_fiji.py
@@ -12,4 +12,4 @@ data_arr = ij.py.to_xarray(img)
 data_arr = data_arr.values.transpose(1, 2, 0).copy()
 phasor = phasor_from_signal(data_arr)
 
-output = ij.py.to_dataset(np.array([phasor[0], phasor[1], phasor[2]]))
+output = ij.py.to_dataset(np.array([phasor[0], phasor[1], phasor[2]], dtype=np.float32))
